backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.8.2

python-jose[cryptography]==3.3.0
email-validator==2.2.0
bcrypt==4.1.2
